Update all scripts to python3

+++ scripts/files.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+#!/usr/bin/env python3
 
 import os
 import shutil
@@ -8,13 +8,13 @@ current_file = os.path.realpath(__file__)
 current_directory = os.path.dirname(current_file)
 cwd = os.getcwd()
 
-print current_file
-print current_directory
-print cwd
+print(current_file)
+print(current_directory)
+print(cwd)
 
 # list files in directory
 for file in os.listdir(current_directory):
-    print file
+    print(file)
 
 # create directory
 new_directory = current_directory + '/' + 'new-directory'
@@ -25,8 +25,8 @@ if not os.path.exists(new_directory):
 # create file
 new_file = new_directory + '/' + 'new-file.txt'
 fo = open(new_file, 'w+')
-fo.write( "Python is a great language.\nYeah its great!!\n");
-print "Name of the file: ", fo.name
+fo.write( "Python is a great language.\nYeah its great!!\n")
+print("Name of the file: ", fo.name)
 fo.close()
 
 # rename a file
@@ -34,19 +34,19 @@ os.rename(new_file, new_directory + '/' + 'new-file2.txt')
 new_file = new_directory + '/' + 'new-file2.txt'
 
 # delete a file
-decide = raw_input("delete the newly created file? ")
-if decide == 'yes':
+decide = input("delete the newly created file? ")
+if decide == 'yes' and os.path.isfile(new_file):
     os.remove(new_file)
-    print new_file + ' deleted'
+    print(new_file + ' deleted')
 
 # delete a directory (only works if directory is empty)
-decide = raw_input("delete the newly created directory? ")
-if decide == 'yes':
+decide = input("delete the newly created directory? ")
+if decide == 'yes' and os.path.exists(new_directory):
     os.rmdir(new_directory)
-    print new_directory + ' deleted'
+    print(new_directory + ' deleted')
 
 # delete a directory (deletes all contents inside)
-decide = raw_input("delete the newly created directory and all files inside? ")
-if decide == 'yes':
+decide = input("delete the newly created directory and all files inside? ")
+if decide == 'yes' and os.path.exists(new_directory):
     shutil.rmtree(new_directory)
-    print new_directory + ' deleted'
+    print(new_directory + ' and all files inside deleted')
